refactor: change VideoValidator to util package
