+++ src/util/VideoValidatorUtil.java
@@ -1,12 +1,11 @@
-package service;
+package util;
 
 import model.Category;
-import util.DateUtil;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
-public abstract class VideoValidator {
+public abstract class VideoValidatorUtil {
     private static final SimpleDateFormat DATE_FORMAT = new SimpleDateFormat("dd/MM/yyyy");
 
     public static void validateTitle(String title) {
